save point

+++ OpenGL_rela/rlglObject.cpp
@@ -2,7 +2,7 @@
 
 using namespace rlgl;
 
-Object::Object(uint32_t _meshID, uint32_t _shaderID, uint32_t _materialID) :
+Object::Object(uint64_t _meshID, uint64_t _shaderID, uint64_t _materialID) :
 	meshID{ _meshID }, shaderID{ _shaderID }, materialID{_materialID}
 {
